+++ UniTAS/Patcher/Services/NoRefresh/INoRefresh.cs
@@ -1,0 +1,6 @@
+namespace UniTAS.Patcher.Services.NoRefresh;
+
+public interface INoRefresh
+{
+    bool Enable { get; set; }
+}