+++ UniTAS/Patcher/Services/UnityEvents/ISceneLoadInvoke.cs
@@ -1,0 +1,6 @@
+namespace UniTAS.Patcher.Services.UnityEvents;
+
+public interface ISceneLoadInvoke
+{
+    void SceneLoadCall();
+}